% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddDeepLcRt.R
\name{add.deeplc.rt}
\alias{add.deeplc.rt}
\title{Adds DeepLC retention times to the predicted Prosit library}
\usage{
add.deeplc.rt(inputLib, outputLib, calibLib, nCal, deepLcFolder)
}
\arguments{
\item{inputLib}{path to spectral library that should have its retention times replaced.}

\item{outputLib}{desired path to spectral library with replaced retention time.}

\item{calibLib}{path to calibration library as created by the MSLibrarian package (.RData)}

\item{nCal}{decimal number giving the proportion of calibration peptides in calibrationLib to use}

\item{deepLcFolder}{path to folder with DeepLC result files}
}
\description{
Adds DeepLC retention times to the predicted Prosit library
}
