% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CeBenchPlot.R
\name{ce.bench.plot}
\alias{ce.bench.plot}
\title{Makes a plot over the best dot products and optimal collision energies based on the spectral matching}
\usage{
ce.bench.plot(ceMat, outLib)
}
\arguments{
\item{ceMat}{input collision energy matrix}

\item{outLib}{path to processed calibration library}
}
\description{
Makes a plot over the best dot products and optimal collision energies based on the spectral matching
}
