% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CheckMsFiles.R
\name{check.ms.files}
\alias{check.ms.files}
\title{Checks if DIA files are valid}
\usage{
check.ms.files(diaFiles)
}
\arguments{
\item{diaFiles}{path to input DIA files}
}
\description{
Checks if DIA files are valid
}
