% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CIRT.R
\docType{data}
\name{cirt}
\alias{cirt}
\title{Conserved iRT peptides}
\format{
An object of class \code{data.frame} with 114 rows and 2 columns.
}
\usage{
data(cirt)
}
\description{
Conserved iRT peptides
}
\keyword{datasets}
