% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateCalibrationLib.R
\name{create.calibration.lib}
\alias{create.calibration.lib}
\title{Creates a calibration library from a set of DIA files and a FASTA file}
\usage{
create.calibration.lib(
  diaFiles,
  fasta,
  projectFolder,
  msConvert = NULL,
  tppDir = NULL,
  openMsDir = NULL,
  filter = "peakPicking true 1-",
  staggeredWindows = F,
  diaUmpireParams = NULL,
  cometParams = NULL,
  spectrastParams = NULL,
  libType = "consensus",
  updateLib = FALSE,
  irt = NULL,
  threads = detectCores()
)
}
\arguments{
\item{diaFiles}{*Input DIA MS files in either RAW or mzXML format}

\item{fasta}{*FASTA file to use for database searching in Comet.}

\item{projectFolder}{*Name of project folder (absolute path). It must be unique.}

\item{msConvert}{path to msConvert executable. If no argument is passed, it will try to autodetect the executable on your system.}

\item{tppDir}{path to Transproteomic pipeline installation folder (for example "C:/TPP"). if no argument is passed, it will try to autodetect the installation folder.}

\item{openMsDir}{path to OpenMS installation folder.(such as "C:/Program Files/OpenMS-2.5.0/"). Will try to autodetect the installation folder if no argument is passed.}

\item{filter}{character vector with msCOnvert filters. Peakpickning on the MS1 level is set as default.}

\item{staggeredWindows}{logical stating if staggered windows were acquired for the input DIA files. (Default is FALSE).}

\item{diaUmpireParams}{*path to a DIA Umpire Signal extraction parameters file. These parameters are necessary for the creation of Pseudo-DDA MS/MS spectra.}

\item{cometParams}{*path to a Comet parameters file necessary for the database searching of pseudo-MS/MS spectra.}

\item{spectrastParams}{*path to a spectrast parameters file with parameters for creation of a spectral library.}

\item{libType}{Type of library to build. Possible values are 1) 'consensus' (default) and 2) 'best_replicate'}

\item{updateLib}{logical indicating if the created library should be updated. Will in this case re-run Spectrast and OpenSwathAssayGenerator to make a new library.}

\item{irt}{Can either be set to "biognosys_irt" or NULL (default). If set to "biognosys_irt", the calibration library retention times will be converted to iRT values.}

\item{threads}{Number of threads or parallel processes to use. Will by default use all available logical processors.}
}
\description{
Creates a calibration library from a set of DIA files and a FASTA file
}
