% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateSpectralLib.R
\name{create.spectral.lib}
\alias{create.spectral.lib}
\title{Creates a spectral library with intensities and retention times from Prosit}
\usage{
create.spectral.lib(
  calibrationLib = NULL,
  projectFolder = NULL,
  msFile = NULL,
  fasta,
  outputLib = NULL,
  format = "spectronaut",
  ceMode = "length_charge",
  threads = detectCores(),
  saveObj = F
)
}
\arguments{
\item{calibrationLib}{path to processed calibration library (RData format)}

\item{projectFolder}{path to project folder containing a processed calibration library.}

\item{msFile}{path to representative pseudo ms/ms file used to create the calibration library (could be extracted from the calibration library file later...)}

\item{fasta}{path to protein sequence FASTA file from Uniprot}

\item{outputLib}{path to desired output library}

\item{format}{output library format c("spectronaut", "openswath")}

\item{ceMode}{Collision energy optimization setting. Possible values are c("length_only", "length_charge", "charge", scalar(fixed collision energy))}

\item{saveObj}{Determines whether a MSLibrarian object will be saved. Possible values are "TRUE" or "FALSE" (default)}
}
\description{
Creates a spectral library with intensities and retention times from Prosit
}
