% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExportSpectraLib.R
\name{export.spectra.lib}
\alias{export.spectra.lib}
\title{Creates a generic spectral library from predicted data, which can be used in OpenSWATH (TSV-format)}
\usage{
export.spectra.lib(msLib, format, calibrationLib, outputLib, batchSize)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{format}{character defining the output library format c("spectronaut", "openswath")}

\item{calibrationLib}{path to calibration library}

\item{outputLib}{character specifying the output library file name (incl. absolute path)}

\item{batchSize}{numeric setting the maximum number of entries to process at a time (to keep the memory usage low)}
}
\description{
Creates a generic spectral library from predicted data, which can be used in OpenSWATH (TSV-format)
}
