% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FeatureFiltering.R
\name{feature.filtering}
\alias{feature.filtering}
\title{Performs a filtering of MSLibrarian-predicted libraries based on extracted features}
\usage{
feature.filtering(inputLib, outputLib, dinoPath, mzml, tolMz, tolRt, threads)
}
\arguments{
\item{inputLib}{input spectral library in Spectronaut .csv format}

\item{outputLib}{filtered output spectral library in Spectronaut .csv format}

\item{dinoPath}{path to the Dinosaur application (.jar) for feature extraction.}

\item{mzml}{path to a mzml file. If a features file is supplied, this parameter is ignored.}

\item{tolMz}{integer specifying the Precursor M/Z tolerance in PPM (default:10)}

\item{tolRt}{allowed retention time tolerance in minutes for targets in library (default:5)}

\item{threads}{number of threads to use for feature matching}
}
\description{
Performs a filtering of MSLibrarian-predicted libraries based on extracted features
}
