% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterLibPrecursors2.R
\name{filter.lib.precursors2}
\alias{filter.lib.precursors2}
\title{Filters slot PrecursorData$Consensus, and returns the filtered Precursor information in PrecursorData$ConsensusFilter}
\usage{
\method{filter}{lib.precursors2}(calibrationLib, exUniMod, precAARange, chargeRange, exAAs)
}
\arguments{
\item{calibrationLib}{input SpectrastLibrary object}

\item{exUniMod}{numeric vector with UniMod accession numbers. A Precursor having any of these modification will be excluded.}

\item{precAARange}{numeric vector specifying the allowed length range in AA, that a Precursor is allowed to have.}

\item{chargeRange}{allowed charge range for precursors.}

\item{exAAs}{character vector describing AAs that a Precursor cannot have in its sequence.}
}
\description{
Filters slot PrecursorData$Consensus, and returns the filtered Precursor information in PrecursorData$ConsensusFilter
}
