% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterPeptides.R
\name{filter.peptides}
\alias{filter.peptides}
\title{Filters peptide sequences based on amino acid length}
\usage{
filter.peptides(
  msLib,
  AALengthRange,
  rejectAAs,
  mzRange,
  chargeRange,
  threads,
  outFile
)
}
\arguments{
\item{msLib}{a scell input object}

\item{AALengthRange}{minimum peptide length allowed}

\item{rejectAAs}{amino acids to reject. Peptides containing these amino acids will be removed.}

\item{mzRange}{mass over charge range to consider for resulting precursor ions}

\item{chargeRange}{charges allowed for resulting precursor ions.}

\item{threads}{number of threads to use for the computation}

\item{outFile}{absolute path to output file for parallel workers}
}
\description{
Filters peptide sequences based on amino acid length
}
