% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterPeptidesTmp.R
\name{filter.peptides.tmp}
\alias{filter.peptides.tmp}
\title{Filters peptide sequences based on amino acid length}
\usage{
\method{filter}{peptides.tmp}(msLib, AALengthRange, rejectAAs, mzRange, chargeRange)
}
\arguments{
\item{msLib}{a scell input object}

\item{AALengthRange}{minimum peptide length allowed}

\item{rejectAAs}{amino acids to reject. Peptides containing these amino acids will be removed.}

\item{mzRange}{mass over charge range to consider for resulting precursor ions}

\item{chargeRange}{charges allowed for resulting precursor ions.}
}
\description{
Filters peptide sequences based on amino acid length
}
