% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterPrego.R
\name{filter.prego}
\alias{filter.prego}
\title{Filter spectral library based on top N high responding peptides for each protein group}
\usage{
\method{filter}{prego}(
  inputLib = NULL,
  outputLib = NULL,
  topPept = NULL,
  pregoPath = NULL,
  replace = F
)
}
\arguments{
\item{inputLib}{absolute path to input spectral library (Spectronaut .CSV format or OpenSwath format .TSV)}

\item{outputLib}{absolute path to the filtered output library}

\item{topPept}{integer giving the top N high responding peptides for each protein group in the library}

\item{pregoPath}{path to the classify.exe tool as part of the PREGO distribution}

\item{replace}{logical determining if the input library should be replaced by the filtered library  (default is FALSE)}
}
\description{
Filter spectral library based on top N high responding peptides for each protein group
}
