% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterProteins.R
\name{filter.proteins}
\alias{filter.proteins}
\title{Subsets a spectral library based on a selection of protein IDs}
\usage{
\method{filter}{proteins}(
  inputLib = NULL,
  outputLib = NULL,
  type = NULL,
  calibrationLib = NULL,
  fdr,
  diaFiles,
  diannPath = NULL,
  replace = F
)
}
\arguments{
\item{inputLib}{path to an input spectral library}

\item{outputLib}{desired path to a subset output spectral library}

\item{type}{means of selecting protein IDs. Possible values are 1) "calibrationlibrary" 2) "diann" or 3) custom character vector of protein IDs.}

\item{calibrationLib}{path to a calibration library object (*.RData). This argument is only used if type == "calibrationlibrary"}

\item{fdr}{numeric decimal value giving the protein group fdr cutoff to use for filtering (default = 0.2, 0 < fdr <= 1).}

\item{diaFiles}{character vector to DIA MS files in .raw format.}

\item{diannPath}{path to DIA-NN executable. Only used if type == "diann". If no path is given, the function will try to auto-detect the DIA-NN installation.}

\item{replace}{should the input library be overwritten/replaced (logical)? Default is FALSE}
}
\description{
Subsets a spectral library based on a selection of protein IDs
}
