% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FilterTransitions.R
\name{filter.transitions}
\alias{filter.transitions}
\title{Performs transition filtering of a library, either based on the top N most intense transitions, or with a defined relative intensity cutoff}
\usage{
\method{filter}{transitions}(
  inputLib = NULL,
  outputLib = NULL,
  topTrans = NULL,
  cutoffTrans = NULL,
  minTrans = NULL,
  replace = F,
  threads = detectCores()
)
}
\arguments{
\item{inputLib}{path to library that will be filtered in Spectronaut (.csv) format.}

\item{outputLib}{path to transition filtered library (Spectronaut *.csv)}

\item{topTrans}{integer specifying the top n transitions to retain for each precursor in the library}

\item{cutoffTrans}{a transition relative intensity cutoff (range 0-1). A cutoff = 0.05 means that transitions having an intensity < 5 % of the maximum intensity (100%) will be removed.}

\item{minTrans}{integer giving the minimum number of transitions that a precursor must have. Default is NULL, meaning that no filtering is performed.}

\item{replace}{logical indicating if the input library be replaced (default = FALSE)}

\item{threads}{number of threads to use for processing}
}
\description{
Performs transition filtering of a library, either based on the top N most intense transitions, or with a defined relative intensity cutoff
}
