% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBestCe.R
\name{get.best.ce}
\alias{get.best.ce}
\title{Extract the optimal collision energy based on peptide length, peptide length and charge, or just precursor charge}
\usage{
get.best.ce(ceMat)
}
\arguments{
\item{ceMat}{input collision energy matrix as a result from running run.spectral.match}
}
\description{
Extract the optimal collision energy based on peptide length, peptide length and charge, or just precursor charge
}
