% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetLibPrecursors2.R
\name{get.lib.precursors2}
\alias{get.lib.precursors2}
\title{Extracts the unique precursor information from an imported calibration library}
\usage{
get.lib.precursors2(calibrationLib)
}
\arguments{
\item{calibrationLib}{input Calibration library object}
}
\description{
Extracts the unique precursor information from an imported calibration library
}
