% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetNoZeroIdx.R
\name{get.no.zero.idx}
\alias{get.no.zero.idx}
\title{Get row indices for transitions in predicted spectral libraries where all intensities are not zero}
\usage{
get.no.zero.idx(files)
}
\arguments{
\item{files}{input spectral library files in generic format (csv)}
}
\description{
Get row indices for transitions in predicted spectral libraries where all intensities are not zero
}
