% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPeptideHist.R
\name{get.peptide.hist}
\alias{get.peptide.hist}
\title{Plots a histogram over peptide masses or precursor m/z values.}
\usage{
get.peptide.hist(sco, massMode, dataSort, bins, xlims)
}
\arguments{
\item{sco}{an input scell object}

\item{massMode}{specifies whether peptide masses or precursor m/z values should be plotted.}

\item{dataSort}{Defines whether masses in slot Peptides or FilterPeptides should be plotted.}

\item{bins}{number of bins to plot}

\item{xlims}{defines the limits for the x-axis}
}
\description{
Plots a histogram over peptide masses or precursor m/z values.
}
