% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrecursorsNew.R
\name{get.precursors.new}
\alias{get.precursors.new}
\title{Generates precursor m/z values for a filtered list of peptides, given the charge range set.}
\usage{
get.precursors.new(msLib, mzRange, chargeRange, matchDb, threads)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{mzRange}{sets the interval for the allowed m/z values.}

\item{chargeRange}{charge ranges for resulting precursor ions}

\item{matchDb}{logical indicating if matching to a prediction database will be performed}

\item{threads}{number of threads to use}
}
\description{
Generates precursor m/z values for a filtered list of peptides, given the charge range set.
}
