% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPredIdx.R
\name{get.pred.idx}
\alias{get.pred.idx}
\title{Get indices for predicted PSMs that match PSMs in a SpectrastLibrary object}
\usage{
get.pred.idx(msLib, threads, outFile)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{threads}{number of processors to use for the parallel computing}

\item{outFile}{absolute path to the output file from the parallel workers}
}
\description{
Get indices for predicted PSMs that match PSMs in a SpectrastLibrary object
}
