% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPredIdxTmp2.R
\name{get.pred.idx.tmp2}
\alias{get.pred.idx.tmp2}
\title{Get indices for predicted PSMs that match PSMs in a SpectrastLibrary object}
\usage{
get.pred.idx.tmp2(msLib, predStr, threads, outFile)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{predStr}{character specifying the selected predicted library to find indices for}

\item{threads}{number of processors to use for the parallel computing}

\item{outFile}{absolute path to the output file from the parallel workers}
}
\description{
Get indices for predicted PSMs that match PSMs in a SpectrastLibrary object
}
