% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPsmIrt.R
\name{get.psm.irt}
\alias{get.psm.irt}
\title{Calculates iRT values for each PSM in psmsPass$SpecLib and plots iRT for predicted vs. experimental}
\usage{
get.psm.irt(pepShkr, prositLib)
}
\arguments{
\item{pepShkr}{input PeptideShaker object}

\item{prositLib}{input predicted spectral library}
}
\description{
Calculates iRT values for each PSM in psmsPass$SpecLib and plots iRT for predicted vs. experimental
}
