% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetSpectralDb.R
\name{get.spectral.db}
\alias{get.spectral.db}
\title{Downloads a Spectral Warehouse SQLite database with predictions of Prosit fragment ion intensities and retention times}
\usage{
get.spectral.db(species = NULL, folder = NULL)
}
\arguments{
\item{species}{species for which a SQLite database should be downloaded. Allowed characters c("H.sapiens", "M.musculus", "S.cerevisiae", "E.coli", "C.elegans", "D.melanogaster")}

\item{folder}{path to download folder. Will default to users Downloads if not specified.}
}
\description{
Downloads a Spectral Warehouse SQLite database with predictions of Prosit fragment ion intensities and retention times
}
