% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HistSpectralMatch.R
\name{hist.spectral.match}
\alias{hist.spectral.match}
\title{Plots histograms of dot products or pearsons correlation coefficientts from spectral matches (experimental vs. predicted)}
\usage{
\method{hist}{spectral.match}(pepShkr, mode, matches, titles, xlab, ylim, col, binSize)
}
\arguments{
\item{pepShkr}{input PepShkr object}

\item{mode}{character specifying whether to plot "dotproduct" or "pearsonsCor"}

\item{matches}{character vector specifying the spectral matches to include in the histograms.}

\item{titles}{character vector specifying the titles for each plotted histogram}

\item{xlab}{character specifying the x-axis label.}

\item{ylim}{numeric vector specifying the range of the y-axis.}

\item{col}{character vector specifying the color of the bars}

\item{binSize}{numeric value giving the number of bins to plot}
}
\description{
Plots histograms of dot products or pearsons correlation coefficientts from spectral matches (experimental vs. predicted)
}
