% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportCalibrationLib.R
\name{import.calibration.lib}
\alias{import.calibration.lib}
\title{Creates a Spectrast Library object, and imports a Spectrast Library which is added to slot OriginalLib}
\usage{
import.calibration.lib(library)
}
\arguments{
\item{library}{absolute path to the unfiltered library file (*.tsv)}
}
\description{
Creates a Spectrast Library object, and imports a Spectrast Library which is added to slot OriginalLib
}
