% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportPrositLibrary.R
\name{import.prosit.library}
\alias{import.prosit.library}
\title{Loads a predicted spectral library and extracts the relevant information such as retention time, relative intensities of fragment ions etc.}
\usage{
import.prosit.library(zipfile, filename)
}
\arguments{
\item{zipfile}{absoute path to the zip file containing the spectral library file}

\item{filename}{name of the file to be extracted from the zip file}
}
\description{
Loads a predicted spectral library and extracts the relevant information such as retention time, relative intensities of fragment ions etc.
}
