% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IrtQcPlot.R
\name{irt.qc.plot}
\alias{irt.qc.plot}
\title{Plots iRT vs. RT with a linear regression. It also returns a summary object of the linear regression}
\usage{
irt.qc.plot(pepShkr)
}
\arguments{
\item{pepShkr}{input PeptideShaker object.}
}
\description{
Plots iRT vs. RT with a linear regression. It also returns a summary object of the linear regression
}
