% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadMs2PipLib.R
\name{load.ms2pip.lib}
\alias{load.ms2pip.lib}
\title{Imports a processed MS2PIP library object into the current MSLibrarian object}
\usage{
load.ms2pip.lib(msLib, ms2pipLib)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{ms2pipLib}{file name of the library to be imported (incl. absolute path)}
}
\description{
Imports a processed MS2PIP library object into the current MSLibrarian object
}
