% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadSpectrastLib.R
\name{load.spectrast.lib}
\alias{load.spectrast.lib}
\title{Loads a Spectrast Library object into the MSLibrarian object}
\usage{
load.spectrast.lib(msLib, file)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{file}{absolute path to Spectrast Library object (saved as RData)}
}
\description{
Loads a Spectrast Library object into the MSLibrarian object
}
