% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeCalibrationLib.R
\name{make.calibration.lib}
\alias{make.calibration.lib}
\title{Create a calibration library in OpenSwath format (TSV)}
\usage{
make.calibration.lib(
  spectrastPath,
  oswGenPath,
  prophetFile,
  libType,
  irt,
  spectrastParams,
  output
)
}
\arguments{
\item{spectrastPath}{absolute path to the Trans-proteomic pipeline binary folder}

\item{oswGenPath}{absolute path to the OpenMS binary folder}

\item{prophetFile}{absolute path to the iProphet results file}

\item{libType}{Type of spectral library to build. Possible values are 1) 'Consensus' and 2) 'Best replicate'"}

\item{irt}{input file with irt values associated to every peptide sequence}

\item{spectrastParams}{input Spectrast parameters file}

\item{output}{Desired output file name of OpenSwath library}
}
\description{
Create a calibration library in OpenSwath format (TSV)
}
