% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeDeepLcInputs.R
\name{make.deeplc.inputs}
\alias{make.deeplc.inputs}
\title{Creates input csv files for retention time predictions in DeepLC}
\usage{
make.deeplc.inputs(inputLib, calibLib, outputFolder, nCal)
}
\arguments{
\item{inputLib}{path to input spectral library in CSV-format (Spectronaut) which should have its retention time changed.}

\item{calibLib}{path to calibration library as created by the MSLibrarian package (.RData)}

\item{outputFolder}{path to output folder where DeepLC input files will be added.}

\item{nCal}{numeric value (nCal > 0 & nCal <= 1), giving the proportion of peptides with known retention times to use for calibration.}
}
\description{
Creates input csv files for retention time predictions in DeepLC
}
