% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMs2pipInputs.R
\name{make.ms2pip.inputs}
\alias{make.ms2pip.inputs}
\title{Creates all the necessary input files for prediction of fragment ion intensities and retention times in MS2PIP (and DeepLC)}
\usage{
make.ms2pip.inputs(
  msLib,
  model,
  fragError,
  libFormat,
  predictData,
  predictRT,
  sampleNbr,
  batchSize,
  threads,
  outputFolder
)
}
\arguments{
\item{msLib}{an input MSLibrarian class object}

\item{model}{character specifying which deeplearning model to use. Available models are c("HCD", "CID", "iTRAQ", "iTRAQphospho", "TMT", "TTOF5600", "HCDch2", "CID")}

\item{fragError}{Fragment ion mass tolerance in Da (only relevant if an MGF file is added)}

\item{libFormat}{character or vector with the desired library file output types, can be c("csv", "mgf", "msp", "spectronaut", "bibliospec")}

\item{predictData}{character indicating if precursors should be predicted from the Sequence object(custom) or the Spectrast object(experimental)}

\item{predictRT}{Logical indicating whether input files should be generated, for retention time predictions in DeepLC.}

\item{sampleNbr}{Integer specifying a random sample number for retention time calibration.}

\item{batchSize}{Integer setting the maximum number of entries of a generated peprec batch file}

\item{threads}{integer specifying the number of threads to use for computation}

\item{outputFolder}{character specifying the absolute path to the folder where all output files will be added.}
}
\description{
Creates all the necessary input files for prediction of fragment ion intensities and retention times in MS2PIP (and DeepLC)
}
