% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeMsms2.R
\name{make.msms2}
\alias{make.msms2}
\title{Adds MS/MS information from Spectral library file into a MSnbase Spectra object (slot: Spectra)}
\usage{
make.msms2(calibrationLib)
}
\arguments{
\item{calibrationLib}{input SpectrastLibrary object}
}
\description{
Adds MS/MS information from Spectral library file into a MSnbase Spectra object (slot: Spectra)
}
