% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeOswLib.R
\name{make.osw.lib}
\alias{make.osw.lib}
\title{Creates a generic spectral library from predicted data, which can be used in OpenSWATH (TSV-format)}
\usage{
make.osw.lib(msLib, selLib, selRT, idSource, libOutput, batchSize)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{selLib}{character setting the spectral library format --> c("PrositLib", "Ms2pipLib")}

\item{selRT}{character defining which indexed retention time predictions to use c("prosit", "ms2pip")}

\item{idSource}{character specifying where the library comes from c("msLibrarian", "spectrast")}

\item{libOutput}{character specifying the output library file name (incl. absolute path)}

\item{batchSize}{numeric setting the maximum number of entries to process at a time (to keep the memory usage low)}
}
\description{
Creates a generic spectral library from predicted data, which can be used in OpenSWATH (TSV-format)
}
