% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePredictionDb.R
\name{make.prediction.db}
\alias{make.prediction.db}
\title{Creates a predicted spectral library database}
\usage{
make.prediction.db(
  prediction_folder,
  task_id,
  sqlite,
  download = T,
  zipped = T
)
}
\arguments{
\item{prediction_folder}{output folder where the predicted spectral library files will be processed and stored}

\item{task_id}{text file with input prosit csv file names and their respective task ids}

\item{sqlite}{preferred name of the SQLite database *.sqlite (It will be added to a subfolder called SQLite in the set prediction_folder)}

\item{download}{should files in the taskid.txt file be downloaded (default = T)}

\item{zipped}{logical indicating whether downloaded files are zipped (default = T)}
}
\description{
Creates a predicted spectral library database
}
