% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePrositCsv.R
\name{make.prosit.csv}
\alias{make.prosit.csv}
\title{Generates a CSV file with precursor information, to be used for prediction of MS/MS spectra in Prosit.}
\usage{
make.prosit.csv(msLib, collisionEngy, chargeRange, outputFile)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{collisionEngy}{Collision energy to use for the prediction in PROSIT.}

\item{chargeRange}{Charge ranges considered for precursor ions.}

\item{outputFile}{full path to the output CSV file}
}
\description{
Generates a CSV file with precursor information, to be used for prediction of MS/MS spectra in Prosit.
}
