% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePseudoSpectra.R
\name{make.pseudo.spectra}
\alias{make.pseudo.spectra}
\title{Creates pseudo DDA spectra f}
\usage{
make.pseudo.spectra(
  msConvertPath,
  mzmlFiles,
  format,
  diaUmpireParams,
  output,
  threads
)
}
\arguments{
\item{msConvertPath}{absolute path to the MsConvert executable}

\item{mzmlFiles}{character vector with absolute paths to input raw files}

\item{format}{character giving the desired converted format}

\item{diaUmpireParams}{character specifying the absolute path to the diaUmpire SE parameters file}

\item{output}{character giving the desired output folder for the converted files}

\item{threads}{integer giving the number of parallel conversion to run}
}
\description{
Creates pseudo DDA spectra f
}
