% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSpectraLib.R
\name{make.spectra.lib}
\alias{make.spectra.lib}
\title{Create a predicted spectral library}
\usage{
make.spectra.lib(msLib, bestCe, predDb, ceMode)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{bestCe}{absolute path to a RData object containing optimal collision energies for MS/MS predictions.}

\item{predDb}{character giving the absolute location of the prediction database.}

\item{ceMode}{character specifying the mode for collision energy optimization. Available modes are c("length_only", "length_charge", "charge", scalar)}
}
\description{
Create a predicted spectral library
}
