% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapDbEntries.R
\name{map.db.entries}
\alias{map.db.entries}
\title{Map entries in the predicted spectral library database to the peptides in the MSLibrarian object}
\usage{
map.db.entries(msLib, predDb, chargeRange)
}
\arguments{
\item{msLib}{input MSLibrarian object.}

\item{predDb}{character giving the absolute location of the prediction database.}

\item{chargeRange}{numeric vector giving the range of precursor charge states to consider.}
}
\description{
Map entries in the predicted spectral library database to the peptides in the MSLibrarian object
}
