% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapPrositPsm.R
\name{map.prosit.psm}
\alias{map.prosit.psm}
\title{Maps indices of Prosit entries to PSMs in a PeptideShaker object.}
\usage{
map.prosit.psm(pepShkr, prositLib)
}
\arguments{
\item{pepShkr}{A PeptideShaker oject}

\item{prositLib}{A processed Prosit library file}
}
\description{
Maps indices of Prosit entries to PSMs in a PeptideShaker object.
}
