% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxTransitions.R
\name{max.transitions}
\alias{max.transitions}
\title{A filter which sets the maximum number of of transitions that a library entry can have}
\usage{
\method{max}{transitions}(msLib, predLib, maxTransitions)
}
\arguments{
\item{msLib}{an input MSLibrarian object}

\item{predLib}{character defining the predicted spectral library to filter c("PrositLib, "Ms2pipLib")}

\item{maxTransitions}{integer setting the maximum number of transitions of a library entry}
}
\description{
A filter which sets the maximum number of of transitions that a library entry can have
}
