% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MergePredictionDbs.R
\name{merge.prediction.dbs}
\alias{merge.prediction.dbs}
\title{Merges prediction databases allowing for creation of hybrid-species spectral libraries}
\usage{
\method{merge}{prediction.dbs}(inputDB, outputDB)
}
\arguments{
\item{inputDB}{character vector with paths to prediction databases that will be merged}

\item{outputDB}{output merged prediction database}
}
\description{
Merges prediction databases allowing for creation of hybrid-species spectral libraries
}
