% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModSpectralLib.R
\name{mod.spectral.lib}
\alias{mod.spectral.lib}
\title{Modifies and filters a spectral library on protein, peptide, or transition level. Also allows for changing library retention times.}
\usage{
mod.spectral.lib(
  projectFolder = NULL,
  inputLib = NULL,
  outputLib = NULL,
  calibrationLib = NULL,
  mods,
  diaFiles,
  protMod,
  protFdr = 0.2,
  diannPath = NULL,
  topPept = NULL,
  pregoPath = NULL,
  cutoffTrans = NULL,
  topTrans = NULL,
  minTrans = NULL,
  nCal = 0.25,
  deepLcPath = NULL
)
}
\arguments{
\item{projectFolder}{path to a MSLibrarian project folder}

\item{inputLib}{path to an input spectral library. If there is only one library, it will be selected automatically if arg - projectFolder is supplied.}

\item{outputLib}{desired path to a modified spectral library. A default name is generated if not specified.}

\item{calibrationLib}{path to a Calibration Library (*.RData format)}

\item{mods}{character vector specifying the modifications to apply. c("protein", "peptide", "transitions", "rt")}

\item{diaFiles}{path to DIA MS raw files used to build the calibration library}

\item{protMod}{type of protein modification to apply. Possible values are 1) "calibrationlibrary", 2) "diann" or 3) a character vector of Uniprot identfiers}

\item{protFdr}{decimal number between 0 - 1 giving the protein group FDR to use for protein ID subsetting of a library.}

\item{diannPath}{path to DIA-NN executable. If not supplied, the function will try to auto-detect the executable on your system.}

\item{topPept}{integer giving the maximum number of high responding peptides (as given by PREGO), to allow for each protein group.}

\item{pregoPath}{path to PREGO executable. If not supplied, the function will try to auto-detect the executable on your system.}

\item{cutoffTrans}{decimal number between 0 - 1 giving the minimum relative intensity that a transition can have (0.05 is default).}

\item{topTrans}{maximum number of high intensity transitions to allow for each precursor in the library. (Is ignored/NULL by default)}

\item{minTrans}{minimum number of transitions to allow for each precursor in a library.}

\item{nCal}{decimal number giving the proportion of calibration library entries to use for RT calibration in DeepLC. (0.25 by default)}

\item{deepLcPath}{path to DeepLC GUI folder. It will try to auto-detect the folder if not supplied.}
}
\description{
Modifies and filters a spectral library on protein, peptide, or transition level. Also allows for changing library retention times.
}
