% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpectralMatch.R
\name{plot.spectral.match}
\alias{plot.spectral.match}
\title{Visualizes spectral matches (experimental vs. predicted) in a mirror plot}
\usage{
\method{plot}{spectral.match}(
  msLib,
  libType,
  specIdx,
  trim,
  mzRange,
  tolType,
  tol,
  precursorRm,
  randIntensity
)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{libType}{character defining the library type c("Consensus","Unfiltered"). Only spectral matches for the selected library type will be plotted.}

\item{specIdx}{numeric vector specifying the spectra indices to plot (if randSpec is FALSE).}

\item{trim}{logical indicating whether spectra should be mz-trimmed or not.}

\item{mzRange}{numeric vector specifying the range of mz-values to keep if trim = TRUE}

\item{tolType}{character specfifying if the tolerance should be "absolute" or "relative".}

\item{tol}{numeric value specifying the tolerance to use}

\item{precursorRm}{logical indicating if the precursor should be removed or not.}

\item{randIntensity}{logical determining if the Prosit relative intensities should be randomly permuted or not}
}
\description{
Visualizes spectral matches (experimental vs. predicted) in a mirror plot
}
