% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotSpectralMatchShkr.R
\name{plot.spectral.match.shkr}
\alias{plot.spectral.match.shkr}
\title{Visualizes spectral matches (experimental vs. predicted) in a mirror plot}
\usage{
plot.spectral.match.shkr(
  pepShkr,
  prositLib,
  indx,
  trim,
  mzRange,
  precursorRm,
  mode,
  toleranceType,
  tolerance,
  randIntensity
)
}
\arguments{
\item{pepShkr}{input PeptideShaker object}

\item{prositLib}{input Prosit spectral library file}

\item{indx}{numeric index speciying the PSM to plot to corresponding predicted spectra.}

\item{trim}{logical indicating if spectra should be mz-trimmed or not.}

\item{mzRange}{numeric vector specifying the range of mz-values to include in the matching.}

\item{precursorRm}{logical indicating if the precursor should be removed or not.}

\item{mode}{character specifying whether "dotProduct" or "cor" should be annotated in the plot}

\item{toleranceType}{character specfifying if the tolerance should be "absolute" or "relative".}

\item{tolerance}{numeric value specifying the tolerance to use}

\item{randIntensity}{logical determining if the Prosit relative intensities should be randomly permuted or not}
}
\description{
Visualizes spectral matches (experimental vs. predicted) in a mirror plot
}
