% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecursorBinning.R
\name{precursor.binning}
\alias{precursor.binning}
\title{Bin precursors based on number cutoff}
\usage{
precursor.binning(ceMat, cutoff)
}
\arguments{
\item{ceMat}{input dot product matrix from predicted vs. experimental spectral matches.}

\item{cutoff}{minimum number of precursors to have a separate bin}
}
\description{
Bin precursors based on number cutoff
}
