% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredAsSpectra.R
\name{pred.as.spectra}
\alias{pred.as.spectra}
\title{Creates a Spectra object (with Spectrum2 objects) for the selected predicted library}
\usage{
pred.as.spectra(msLib, selPred, threads, outFile)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{selPred}{character specifying the predicted library to use c("Prosit", "MS2PIP")}

\item{threads}{number of processors to use for the parallel computation}

\item{outFile}{absolute path to output text file with console info from parallel workers}
}
\description{
Creates a Spectra object (with Spectrum2 objects) for the selected predicted library
}
