% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessCalibrationLib.R
\name{process.calibration.lib}
\alias{process.calibration.lib}
\title{Creates and processes a calibration library object}
\usage{
process.calibration.lib(
  specLib = NULL,
  projectFolder = NULL,
  predictionDb,
  rt = NULL,
  outLib = str_replace(specLib, ".tsv", ".RData"),
  threads = detectCores()
)
}
\arguments{
\item{specLib}{path to spectral library (in TSV-format, as processed by make.calibration.lib). Not necessary if the project Folder argument is added}

\item{projectFolder}{Path to project folder. If not specified it is necessary to add the specLib argument.}

\item{predictionDb}{path to SQLite database with Prosit predictions.}

\item{rt}{character specifying the input library retention time format. Possible values: c("seconds", "iRT")}

\item{outLib}{desired path to the output calibration library (.RData format)}

\item{threads}{number of  threads to use}
}
\description{
Creates and processes a calibration library object
}
