% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessMs2PipLibrary.R
\name{process.ms2pip.lib}
\alias{process.ms2pip.lib}
\title{Processes a MS2PIP predicted library and saves it as a RData object}
\usage{
process.ms2pip.lib(zipfilePred, zipfileRT, output)
}
\arguments{
\item{zipfilePred}{The absolute path to the zipped intensity predictions output file from MS2PIP (in spectronaut.csv format)}

\item{zipfileRT}{The absolute path to the zipped retention time predictions output file from DeepLC (in spectronaut.csv format)}

\item{output}{Output file name for the RData object (including absolute path)}
}
\description{
Processes a MS2PIP predicted library and saves it as a RData object
}
