% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessPrositLibrary.R
\name{process.prosit.lib}
\alias{process.prosit.lib}
\title{Loads a predicted spectral library and extracts the relevant information such as retention time, relative intensities of fragment ions etc.}
\usage{
process.prosit.lib(
  zipfile,
  noZeroIntensity,
  transitionFilter,
  transitionThreshold,
  output
)
}
\arguments{
\item{zipfile}{absoute path to the zip file containing the spectral library file}

\item{noZeroIntensity}{logical indicating if all transitions with zero intensities should be removed}

\item{transitionFilter}{logical indicating if a transition filter should be used}

\item{transitionThreshold}{numeric value (0,1) setting the percentage of most observed transitions to include}

\item{output}{absolute path to processed library file, as a RData object}
}
\description{
Loads a predicted spectral library and extracts the relevant information such as retention time, relative intensities of fragment ions etc.
}
