% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prositAsSpectra.R
\name{prosit.as.spectra}
\alias{prosit.as.spectra}
\title{Make part of prosit library file into Spectrum2 objects contained in a Spectra object}
\usage{
prosit.as.spectra(msLib, threads, outFile)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{threads}{number of processors to use for the parallel computation}

\item{outFile}{absolute path to output text file with console info from parallel workers}
}
\description{
Make part of prosit library file into Spectrum2 objects contained in a Spectra object
}
