% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prositAsSpectrum2.R
\name{prosit.as.spectrum2}
\alias{prosit.as.spectrum2}
\title{Make part of prosit library file into Spectrum2 objects contained in a Spectra object}
\usage{
prosit.as.spectrum2(pepShkr, prositLib)
}
\arguments{
\item{pepShkr}{input PepShkr object}

\item{prositLib}{input prosit library file}
}
\description{
Make part of prosit library file into Spectrum2 objects contained in a Spectra object
}
