% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProteinFilter.R
\name{protein.filter}
\alias{protein.filter}
\title{Subsets an input spectral library based on protein IDs}
\usage{
protein.filter(inputLib, outputLib, proteinIds)
}
\arguments{
\item{inputLib}{path to input spectral library to be filtered}

\item{outputLib}{desired path to subsetted spectral library}

\item{proteinIds}{character vector of protein IDs which the subsetted library must contain}
}
\description{
Subsets an input spectral library based on protein IDs
}
