% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadFasta.R
\name{read.fasta}
\alias{read.fasta}
\title{Reads a fasta file of AA sequences and stores the information in slot Proteins of a MSLibrarian Sequence object.}
\usage{
read.fasta(file)
}
\arguments{
\item{file}{A complete path to a fasta file of AA sequences}
}
\description{
Reads a fasta file of AA sequences and stores the information in slot Proteins of a MSLibrarian Sequence object.
}
