% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReplaceRts.R
\name{replace.rts}
\alias{replace.rts}
\title{Replaces the retention times for an input spectral library}
\usage{
replace.rts(
  inputLib = NULL,
  outputLib = NULL,
  calibrationLib = NULL,
  outputFolder = NULL,
  nCal = 0.33,
  deeplc = NULL,
  replace = F,
  threads = detectCores()
)
}
\arguments{
\item{inputLib}{input spectral library that will have its retention times replaced.}

\item{outputLib}{path to output library}

\item{calibrationLib}{calibration library object used to extract calibration peptides for DeepLC}

\item{outputFolder}{desired path to output folder for DeepLC results}

\item{nCal}{decimal number giving the proportion of calibration peptides in calibrationLib to use (default = 0.33)}

\item{deeplc}{path to the DeepLC GUI installation folder (will try to auto-detect it by default)}

\item{replace}{should the input library be replaced(logical)? Default is FALSE}
}
\description{
Replaces the retention times for an input spectral library
}
