% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunComet.R
\name{run.comet}
\alias{run.comet}
\title{Runs the database search engine Comet on input spectral files}
\usage{
run.comet(cometPath, msFiles, fasta, cometParams, threads)
}
\arguments{
\item{cometPath}{character giving the absolute path to the Comet executable}

\item{msFiles}{absolute path to the input MS files (mzXML)}

\item{fasta}{character giving the absolute path to the desired FASTA file}

\item{cometParams}{character giving the absolute path to the Comet parameters file}

\item{threads}{integer specifying the number of threads to use for the conversion}
}
\description{
Runs the database search engine Comet on input spectral files
}
