% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDeepLc.R
\name{run.deeplc}
\alias{run.deeplc}
\title{Runs DeepLC to predict retention times for a set of peptides}
\usage{
run.deeplc(
  inputFolder,
  splitCal = 50,
  minDivider = 10,
  maxDivider = 5000,
  deeplc,
  threads
)
}
\arguments{
\item{inputFolder}{path to folder with input files}

\item{splitCal}{split size for predictions}

\item{minDivider}{minimum dictionary divider}

\item{maxDivider}{maximum dictionary divider}

\item{deeplc}{path to deeplc_gui folder}

\item{threads}{number of threads to use for predictions}
}
\description{
Runs DeepLC to predict retention times for a set of peptides
}
