% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMsConvert.R
\name{run.msconvert}
\alias{run.msconvert}
\title{Converts MS RAW files to desired output format using MsConvert}
\usage{
run.msconvert(msConvertPath = NULL, rawFiles, format, filter, output, threads)
}
\arguments{
\item{msConvertPath}{absolute path to the MsConvert executable}

\item{rawFiles}{character vector with absolute paths to input raw files}

\item{format}{character giving the desired converted format}

\item{filter}{character specifing the filter to use during the conversion}

\item{output}{character giving the desired output folder for the converted files}

\item{threads}{integer giving the number of parallel conversion to run}
}
\description{
Converts MS RAW files to desired output format using MsConvert
}
