% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunMsFragger.R
\name{run.msfragger}
\alias{run.msfragger}
\title{Runs the MSFragger database search tool on pseudo-DDA MS/MS spectra}
\usage{
run.msfragger(msfragger, fragParams, msFiles, fasta, decoyDb)
}
\arguments{
\item{msfragger}{path to MSFragger executable}

\item{fragParams}{path to MSFragger parameters file}

\item{msFiles}{input pseudo-dda MS/MS files}

\item{fasta}{input fasta file}

\item{decoyDb}{path to DecoyDatabase executable (part of OpenMS)}
}
\description{
Runs the MSFragger database search tool on pseudo-DDA MS/MS spectra
}
