% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunProphets.R
\name{run.prophets}
\alias{run.prophets}
\title{Runs PeptideProphet and iProphet on input Comet search engine results}
\usage{
run.prophets(
  interactPath,
  peptideProphetPath,
  interProphetPath,
  searchFiles,
  output,
  threads
)
}
\arguments{
\item{interactPath}{Absolute path to the interactparser.exe}

\item{peptideProphetPath}{Absolute path to the peptideprophetparser.exe}

\item{interProphetPath}{Absolute path to the interprophetparser.exe}

\item{searchFiles}{input Comet database search result files (pepXML)}

\item{output}{Absolute path to the desired output folder}

\item{threads}{integer giving the number of threads to use}
}
\description{
Runs PeptideProphet and iProphet on input Comet search engine results
}
