% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSpectralMatch2.R
\name{run.spectral.match2}
\alias{run.spectral.match2}
\title{Runs spectral comparisons between experimental spectra in the calibration library object and predicted spectra}
\usage{
run.spectral.match2(spectra, precursors, predictionDb, threads)
}
\arguments{
\item{spectra}{input Spectrum2 objects of Calibration Library object}

\item{precursors}{filtered precursor data in Calibration library object}

\item{predictionDb}{absolute path to the prediction database}

\item{threads}{integer giving the number of threads to use for computation}
}
\description{
Runs spectral comparisons between experimental spectra in the calibration library object and predicted spectra
}
