% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScatterSpectralMatch.R
\name{scatter.spectral.match}
\alias{scatter.spectral.match}
\title{Draws a scatter plot of either dotproducts or pearsons correlation coefficients for spectral matches.}
\usage{
scatter.spectral.match(pepShkr, mode, matches, xlab, ylab, main)
}
\arguments{
\item{pepShkr}{input PeptideShaker object}

\item{mode}{character specifying if dot products or pearsons correlation coefficient should be plotted.}

\item{matches}{character specifying which spectral matches to plot against each other(x vs. y)}

\item{xlab}{character specifying the x-axis label}

\item{ylab}{character specifying the y-axis label}

\item{main}{character specifying the title of the plot}
}
\description{
Draws a scatter plot of either dotproducts or pearsons correlation coefficients for spectral matches.
}
