% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralMatchShkr.R
\name{spectral.match.shkr}
\alias{spectral.match.shkr}
\title{Returns dot-product, pearsons correlation and common peaks for matches between PSMs in PepShkr object and predicted spectra from Prosit.}
\usage{
spectral.match.shkr(
  pepShkr,
  prositLib,
  toleranceType,
  tol,
  trim,
  mzRange,
  listName,
  precursorRm,
  randIntensity
)
}
\arguments{
\item{pepShkr}{input pepShkr object}

\item{prositLib}{input spectral library file from Prosit}

\item{toleranceType}{A character specifying if matching peak tolerance should be "relative" or "absolute"}

\item{tol}{numeric specifying the matching tolerance}

\item{trim}{logical indicating if trim should be performed or not}

\item{mzRange}{numeric vector indicating mz values to be removed}

\item{listName}{Name of list to add results to}

\item{precursorRm}{logical indicating whether to remove intense precursors to improve spectral matching}

\item{randIntensity}{logical indicating if random permutation of intensities in Prosit Spectrum2 obj. should be performed.}
}
\description{
Returns dot-product, pearsons correlation and common peaks for matches between PSMs in PepShkr object and predicted spectra from Prosit.
}
