% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpectralMatchTmp.R
\name{spectral.match.tmp}
\alias{spectral.match.tmp}
\title{Compares experimental spectra (Spectrast) with corresponding predicted spectra (for example Prosit)}
\usage{
spectral.match.tmp(
  msLib,
  predLib,
  tolType,
  tol,
  trim,
  mzRange,
  precursorRm,
  randIntensity,
  listName,
  threads,
  outFile
)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{predLib}{character defining the spectral library to use for prediction}

\item{tolType}{A character specifying if matching peak tolerance should be "relative" or "absolute"}

\item{tol}{numeric specifying the matching tolerance}

\item{trim}{logical indicating if trim should be performed or not}

\item{mzRange}{numeric vector indicating mz values to be removed}

\item{precursorRm}{logical indicating whether to remove intense precursors to improve spectral matching}

\item{randIntensity}{logical indicating if random permutation of intensities in Prosit Spectrum2 obj. should be performed.}

\item{listName}{character specifying the list name for accessing the metrics of the spectra comparison}

\item{threads}{number of processors to use for the parallel processing}

\item{outFile}{absolute path to the output text file with console info from the parallel workers}
}
\description{
Compares experimental spectra (Spectrast) with corresponding predicted spectra (for example Prosit)
}
