% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetPrecursors.R
\name{get.precursors}
\alias{get.precursors}
\title{Generates precursor m/z values for a filtered list of peptides, given the charge range set.}
\usage{
get.precursors(msLib, mzRange, chargeRange, threads, outFile)
}
\arguments{
\item{msLib}{a scell object}

\item{mzRange}{sets the interval for the allowed m/z values.}

\item{chargeRange}{charge ranges for resulting precursor ions}

\item{threads}{number of threads to use for the computation}

\item{outFile}{absolute path to output file with console information from parallel workers}
}
\description{
Generates precursor m/z values for a filtered list of peptides, given the charge range set.
}
