% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadPrositLibrary.R
\name{load.prosit.lib}
\alias{load.prosit.lib}
\title{Loads a processed Prosit Library file into a MSLibrarian object}
\usage{
load.prosit.lib(msLib, file)
}
\arguments{
\item{msLib}{input MSLibrarian object}

\item{file}{absolute path to processed Prosit library file}
}
\description{
Loads a processed Prosit Library file into a MSLibrarian object
}
