% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakePrositCsv.R
\name{make.prosit.csv}
\alias{make.prosit.csv}
\title{Generates a CSV file with precursor information, to be used for prediction of MS/MS spectra in Prosit.}
\usage{
make.prosit.csv(fasta, chargeRange, ceRange, prefix, outputFolder, threads)
}
\arguments{
\item{fasta}{input fasta file}

\item{chargeRange}{Charge ranges considered for precursor ions.}

\item{ceRange}{Collision energy range for predictions.}

\item{prefix}{file prefix to use (like the species from which the fasta file is derived)}

\item{outputFolder}{full path to the output CSV file}

\item{threads}{number of threads to use}
}
\description{
Generates a CSV file with precursor information, to be used for prediction of MS/MS spectra in Prosit.
}
