% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessPrositLibrary.R
\name{process.prosit.lib}
\alias{process.prosit.lib}
\title{Loads a predicted spectral library and extracts the relevant information such as retention time, relative intensities of fragment ions etc.}
\usage{
process.prosit.lib(zipfile, filename, output)
}
\arguments{
\item{zipfile}{absoute path to the zip file containing the spectral library file}

\item{filename}{name of the file to be extracted from the zip file}

\item{output}{absolute path to processed library file, as a RData object}
}
\description{
Loads a predicted spectral library and extracts the relevant information such as retention time, relative intensities of fragment ions etc.
}
