% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunDiaNN.R
\name{run.diann}
\alias{run.diann}
\title{Runs a DIA-NN analysis of the selected DIA-file and input library in TSV format}
\usage{
run.diann(
  diannPath,
  diaFile,
  libFile,
  output = NULL,
  matrices = T,
  matQ = 0.01,
  threads = detectCores(),
  fasta = NULL,
  enzyme,
  separateMassAcc = F,
  robustLC = NULL
)
}
\arguments{
\item{diannPath}{Absolute path to the DIA-NN executable}

\item{diaFile}{Absolute path to the input DIA raw file}

\item{libFile}{Absolute path to the input spectral library file (as given by the make.osw.lib function)}

\item{output}{Desired file name for DIA-NN reports (incl. absolute path)}

\item{matrices}{A logical determining if quantitative matrices for proteins and precursors should be outputted (Default=TRUE).}

\item{matQ}{numeric in the range c(0,1) setting the q-value to use for filtering matrices.}

\item{threads}{Integer specifying the number of threads for the analysis}

\item{fasta}{Absolute path to a FASTA file for matching of peptides sequences to proteins}

\item{enzyme}{Character specifying the enzyme to use for in-silico digestion of sequences in the FASTA file}

\item{separateMassAcc}{should mass accuracy be determined for each individual MS file (logical)}

\item{robustLC}{Specfies whether Quantification should be run in Robust LC mode. Possible values are "high_accuracy" or "high_precision". Default = NULL.}
}
\description{
Runs a DIA-NN analysis of the selected DIA-file and input library in TSV format
}
